requirements.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
fastapi==0.70.0
h11==0.12.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
numpy==1.21.4
pandas==1.3.4
protobuf==3.19.1
pydantic==1.8.2
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==4.0.1
uvicorn==0.15.0
